backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/ML
langchain-google-genai
google-generativeai
openai==1.3.5
langchain==0.0.335
sentence-transformers==2.2.2
spacy==3.7.2
transformers==4.35.0

# Vector Database
pinecone-client==2.2.4

# Graph Database
neo4j==5.14.1

# Data Processing
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# Database
redis==5.0.1
asyncpg==0.29.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Utilities
pydantic==2.5.0
httpx==0.25.1
pydantic-settings
